College Project/final_year_project/collection/FinalYearProject8/PR3/requirements.txt
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
mysql-connector==2.2.9
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
